CHATBOT-123: Devolver campos para que se puedan generar los mensajes personalizados por tipo de resultado de reserva (MontoMinimo, MOnotMaximo, etc).

+++ app/Portal.Consultoras.Web/Models/IngresoExternoModel.cs
@@ -8,7 +8,7 @@
         public string Version { get; set; }
         public string Campania { get; set; }
         public string NumeroPedido { get; set; }
-        public int ProcesoId { get; set; }
+        public long ProcesoId { get; set; }
         public string AutoReservar { get; set; }
     }
 }
